bridge: Require result conversion to only produce BridgeLayerErrors
Similar to the previous commit, this means they can't affect the
checked exception signature of the associated method (apart from
manually throwing exceptions, or invoking callbacks in the middle of
argument conversion).

This requires tweaking the macro-provided body of bridge_fns a bit, to
not blithely pass errors through jni::ResultTypeInfo. The
functionality should not change, however.