zkcredential: Keep compressed *and* decompressed endorsements around
This saves work for callers that need both, which includes
GroupSendEndorsement: after receiving and validating the endorsements,
they need to get serialized and sent back up to the app layer to put
in its database (compressed), but we also generate an extra
"everyone-but-me" endorsement from the results (decompressed).

This saves quite a bit of time in the app-layer benchmarks, since they
include the cost of serialization.