Throw JS deserialization errors of correct type
Bind the `this` value for the deserialization check callbacks when invoking 
them so that the Rust code can access properties on the Native JS module. 
Without this, the Rust code is called with a "global object" as the context, 
which means it fails to construct a LibSignalErrorBase and has to fall back to 
an untyped error message.