swift: Add ConnectionEventsListener for UnauthenticatedChatService
This follows Node in splitting out a base protocol for "disconnected"
events that the existing ChatListener protocol extends. It's a bit
more involved because of the helper class that keeps track of both the
listener and the ChatService, which can't be made generic because
Swift generics aren't monomorphized (and thus won't work as C function
pointers).

This is a breaking change in practice because the name of the callback
has changed (from chatServiceConnectionWasInterrupted to just
connectionWasInterrupted).