CI: Don't bother building x86_64 Android for every PR
We don't run any "instrumented tests" for Android, so we're mostly
only building in case there are Android-specific issues. We could
consider eliminating one of the ARM slices as well, but Android is the
only target that uses 32-bit ARM, and 64-bit ARM is the dominant
variant (and what we measure the code size of).