node: Make downcasting libsignal errors a little more convenient
And add missing cases to the LibSignalError enum.

All libsignal errors (nearly) are instances of a common
LibSignalErrorBase, differentiated by an enum `code`. Certain codes
indicate the presence of additional properties. By exposing a type
predicate method `is` that checks the code, we can get TypeScript to
understand which additional properties might be present.

Before:

    if (error.code === ErrorCode.InvalidRegistrationId) {
      const error2 = error as InvalidRegistrationIdError;
      use(error2.addr());
    }

After:

    if (error.is(ErrorCode.InvalidRegistrationId)) {
      use(error.addr());
    }