backup: Pipeline backup processing
Allow the protobuf parsing and validation of frames to happen in
parallel with the reading and splitting *into* frames from a backup
file, particularly an *encrypted* backup file. Pays the fixed cost of
spawning a thread in order to make things better for larger backups.
On a Pixel 2, this provides almost a textbook pipelining speedup of
25-30%. (There are other possible split points in the pipeline, but
this one is the most straightforward.)

Note that there were a lot of finicky choices here; in particular,
using blocking recv() instead of a manual sleep loop barely showed any
improvement at all, mostly because the producer thread was spending
too much time waking up the consumer thread.