backup: Factor out key-parsing logic for tools
This allows decrypt_backup and encrypt_backup to use the full set of
options for specifying key material as the main validator CLI, though
they will continue to use a fallback AEP and ACI if no key material is
provided. The upcoming scrambler tool will likewise use the same
options for decrypting.