Sink EC key types from libsignal_protocol to libsignal_core::curve
This is necessarily a breaking change because the EC key operations
can no longer return SignalProtocolError. However, the types are still
exported through libsignal_protocol, and SignalProtocolError
implements From<CurveError>, so hopefully any clients will be guided
to update to the new usage.

Other changes:

- PublicKey::public_key_bytes no longer uses Result; any future curve
  types will also need to keep their public key in a borrowable form.

- PublicKey::verify_signature also no longer uses Result; if a future
  curve type does not support verifying a signature, or if the
  signature is structurally invalid, that should be treated the same
  as a verification failure. (This is already how structurally invalid
  Curve25519 signatures were being treated.)

This change allows other crates to depend on the smaller
libsignal_core rather than all of libsignal_protocol (though it also
means libsignal_core is no longer as small as it used to be).