backup: Aggressively push uncommon message data out of line
A backup has a *lot* of ChatItemData instances, which get stored
inline in a Vec. So we benefit from them not being too large, and from
not having drastically different sizes in the enum. At the same time,
we want to avoid having many allocations for the most *common* message
types.

The biggest offender here is attachments, which have a lot of
potential (nested) fields that are usually present. And unfortunately,
StickerMessage and VoiceMessage are both attachment-based. For those
we'll accept the cost of out-of-line allocations, as with
ContactMessage.

Other changes are likely to be a good bet at most times:

- Standard messages may or may not have quotes, and almost never have
  long text.

- Payment messages are likely rare enough that we shouldn't optimize
  for them with inline storage.

For the artifical test case in the validation benchmark, this is a big
speedup (~10%), since all messages are StandardMessages without long
text.