net: Thread a "log tag" down into ws2/chat
This is mainly important for chat connections, which are often paired
auth and unauth, and thus result in two very similar log lines right
next to each other. For now the tags are just the constant strings
"auth", "unauth", and "cdsi", plus some hardcoded values for the old
TransportConnector interfaces (and examples and tests); we may want to
refine them in the future.