Downgrade pqcrypto-internals back to 0.2.5
The changes in 0.2.6-8 all concern adding support for performing two
SHA-3 operations in parallel, including an AArch64 assembly version
that causes problems for some of our compiler versions pinned in
Docker. This operation isn't (currently) used in pqcrypto-kyber, and
there aren't any other changes we're missing out on, so this downgrade
should be safe.