net: Introduce ConnectorFactory and WebSocketTransportConnectorFactory
Formalizes the `Fn() -> Connector` introduced in 02653167 into a trait
of its own, ConnectorFactory, so that nameable types with state can
implement it (rather than just closures and function pointers).
Genericizes several APIs to work with multiple kinds of
ConnectorFactory and not just the default one. This will be useful in
the future.

WebSocketTransportConnectorFactory is effectively serving as an alias
for a ConnectorFactory suitable for connecting websockets over; it
implies a handful of constraints that would normally be broken up into
multiple lines. It's not strictly necessary but it does make the code
clearer.