net: Factor common parameters out as ConnectionResources
In practice, ConnectState is used for websocket connections, which
means it will (nearly) always be used with a DNS resolver, a network
change event, and an optional confirmation header to look for. Group
these four values (including the lock-protected state itself) in a new
type, which can also then serve as the `self` type for the connection
operations.