net: Check that TCP stream writes complete sensibly on Macs
Background: there is at least one VPN where if you use it on a Mac,
the kernel write() call will return *the socket file descriptor*
rather than the number of bytes written if the connection is over
IPv6. This value previously propagated into BoringSSL, which detected
it in routine buffer management code as an overflow, and, being
written in C, had no recourse but to abort the program. Now we produce
a write error instead, which successfully gets reported as a
std::io::Error in Rust.