backup: Assert that we don't lose any data unwrapping a reader
VarintDelimitedReader has an internal buffer; that buffer is *nearly*
always drained after reading a whole frame, but a sufficiently small
frame would fail to do so. This is fine if the caller continues
reading frames, but could lose those bytes if the inner reader is
extracted. We don't expect this to be a problem in any real scenarios,
but we'd like to know if it ever is.