ffi: Convert *most* error bridging to a simpler trait
The new IntoFfiError allows most errors to forego individual
implementation of FfiError's requirements and just eagerly convert to
the smallest bridged form. This is simpler for common cases, and also
makes it easier to have cases where you *do* delegate to another
error's implementation, since you don't have to do it callback by
callback. It *does* mean stringifying the message eagerly more often,
but in practice the Swift side of the bridge nearly always asks for it
anyway; in the cases where it doesn't, we still have the more flexible
FfiError trait when it matters.

This initial commit only changes over all the easy cases and is
*nearly* identical behavior to what was there before (one message
changed). We should be able to go further as well with more
helper types.