net: Remove LogSafeTungsteniteError in favor of WebSocketServiceError
Both of these types simplify tungstenite::Error, but
LogSafeTungsteniteError was designed for *late* conversion from a
tungstenite::Error (just before logging or wrapping in IoError), while
WebSocketServiceError threw away more (but not all) information
eagerly, and also has an extra case for a libsignal-net-initiated
timeout closure. It's not a 100% overlap, but the 90% overlap that
*was* there was inconvenient and a bit confusing.
WebSocketServiceError is good enough for how LogSafeTungsteniteError
was used.