net: ComposedConnector can always use the outer connector's Error
Previously, ComposedConnector had an error type as a generic argument,
which its inner and outer connectors had to be compatible with. In
practice, though, all of the error types for the outer connectors we
use are already errors that include a case for the inner connector
failing, so we can get rid of the flexibility of the generic argument
and make our use sites simpler.