net: Add http_version to HttpRouteFragment and ConnectionConfig
Setting this at the route level is a requirement that doesn't quite
match up with the "negotiation" part of ALPN. However, we currently
know exactly which HTTP version we'll use to connect to any given
server: HTTP/1.1, unless we specifically know in advance the server
supports H2. That's sufficient for our uses here.