bridge: Use bridge_callbacks for PreKeyStore (Java bridge)
Adds support for callbacks with results to the JNI bridge, similar to
what f2eafbe6f8 did for the C bridge.

This does require introducing an "internal" version of PreKeyStore
that has simpler signatures for its callbacks; previously, the Java
objects for bridge_handle types were constructed on the Rust side of
the bridge, but that's not how bridge_fn works, and so it's not how
bridge_callbacks works. The additional overhead should not be
prohibitive.